nixos-rebuild-ng: move test to the correct file
